feat: honour "expose" property of http-errors
This commit adds support for the "expose" property used by the
http-errors module, which allows developers to explicitly signal the
error message to be passed to the HTTP response for errors. This is
useful for certain errors (e.g., those with status code >=500) where the
original error message is otherwise suppressed by default.

Resolves: #78
Signed-off-by: getsnoopy <youcangetsnoopy+git@gmail.com>